// src/model/UserState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ChannelId } from "./ChannelId";
import type { PostId } from "./PostId";
import type { UserId } from "./UserId";

export type UserState = { id: UserId, name: string, subscriptions: { [key in ChannelId]?: Array<PostId> }, };