// src/model/Post.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ChannelId } from "./ChannelId";
import type { PostId } from "./PostId";

export type Post = { id: PostId, channelId: ChannelId, title: string, link: string, thumbnailUrl: string, summary: string, lastCommentTime: string | null, };